+++ cli/src/test-utils/test-runner.spec.ts
@@ -1,0 +1,1 @@
+describe("test runner", () => {});
